chore: split menut toos
