+++ webapp/pages/topics/tools-menu.tsx
@@ -1,9 +1,9 @@
 import { ActionIcon, Text, Menu } from "@mantine/core";
 import { IconInfoCircle, IconTool, IconTrash } from "@tabler/icons";
-import { deleteTopic } from "../../../tauri/admin";
+import { deleteTopic } from "../../tauri/admin";
 import { useNavigate } from "react-router-dom";
 import { openConfirmModal } from "@mantine/modals";
-import { useNotifications } from "../../../providers";
+import { useNotifications } from "../../providers";
 
 export const ToolsMenu = ({ clusterId, topic }: { clusterId: string; topic: string }) => {
   const navigate = useNavigate();
