+++ webapp/pages/topics/topic/main.tsx
@@ -2,15 +2,15 @@ import { Center, Container, Loader } from "@mantine/core";
 import { RecordsList, RecordsListRef } from "./record-list";
 import { getConsumerState, stopConsumer } from "../../../tauri/consumer";
 import { PageHeader } from "../../../components";
-import { openConsumerModal } from "./consumer-modal";
+import { openConsumerModal } from "./modals/consumer-modal";
 import { useQuery } from "@tanstack/react-query";
 import { getLastOffsets, getTopicInfo } from "../../../tauri/admin";
 import { Allotment } from "allotment";
-import { ToolsMenu } from "./tools-menu";
+import { ToolsMenu } from "../tools-menu";
 import { TopicPageMenu } from "./topic-menu";
 import { useRef, useState } from "react";
 import { useCache } from "../../../hooks";
-import { ExportRecordsModal } from "./export-records-modal";
+import { ExportRecordsModal } from "./modals/export-records-modal";
 
 export const Topic = ({ clusterId, topicName }: { clusterId: string; topicName: string }) => {
   // cached query across navigation
