+++ webapp/pages/topics/topic/modals/export-records-modal.tsx
@@ -1,8 +1,8 @@
 import { Text, Checkbox, Modal, Title, NumberInput, Stack, Group, Button } from "@mantine/core";
-import { useNotifications } from "../../../providers";
+import { useNotifications } from "../../../../providers";
 import { save } from "@tauri-apps/api/dialog";
 import { useState } from "react";
-import { exportRecords } from "../../../tauri/consumer";
+import { exportRecords } from "../../../../tauri/consumer";
 
 type ExportRecordsModalProps = {
   clusterId: string;
